requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.10
python-dotenv>=1.0.0

# HTTP Client for MCP API calls
httpx>=0.25.0